+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - React 17.
 - `CachedDataComponent` component.
 - `SameSite` flag in cookies.
+- Option to register http method in route callbacks in `Server.registerRouteCallback`.
 ### Breaking changes
 - Removed `Button` component.
 - Removed `res.render` function.